Add instructions to run on mobile
The Media Capabilities API requires the application to be serverd over
HTTPS. This is because the API exposes information about the user's
device, and the browser wants to ensure that this information is not
intercepted by a third party.

To run this example on a mobile device, you will need to serve the
application over HTTPS. You can do this by running a local server
with a self-signed certificate. The added instructions will
walk developers through this process.